Project Plan
Higher Diploma in Data Analytics GMIT
   Programming and Scripting
   PANDS Project

A: Project Introduction
This project has been prepared to show the power of the Python programming language in interpreting a dataset. The sample dataset to be interpreted is Fisher’s Iris Dataset. 
Ronald Fisher  (1890 - 1962) was a biologist who through his work on population genetics invented statistical hypothesis testing and design of experiments (DOE). The Iris dataset was published and discussed in his 1936 paper The use of multiple measurements in taxonomic problems . The data was collected to quantify the physical variation of Iris flowers of three related species. The data set consists of 50 samples from all three species of iris (Iris setosa, Iris virginica, Iris versicolor). The four features measured were the length and width of the sepals and petals in centimetres. The combination of these four features can be used to identify each particular species.
Python is a computing programming language which is high-level and general-purpose. It was created in 1991 by Guido van Rossum with the aim of emphasising code readability by being user-friendly and easy to learn   .
B: Project Plan
This project will be completed in a number of steps. 
First research will be performed via an internet search on the Iris dataset. This aim will be to find the dataset, investigate techniques for manipulating and formatting the dataset, and research what work has previously been performed on the dataset that can be incorporate into this project. 
Next, the project will be broken down into several simple modules to allow the project to easily managed in small parts. For each module the input, process and output will be identified. Research will also be performed to find code that can be used to get meaningful output from the data.
Then, using the formatted dataset, code will be developed to evaluate the dataset and provide an output. The result of this output will be discussed. Conclusions on the output and the work performed will be discussed.

C: Research
A quick internet search was performed for research purposes on the topics discussed below.
a.	Fisher’s Iris Data Set
         https://en.wikipedia.org/wiki/Ronald_Fisher
         https://en.wikipedia.org/wiki/Iris_flower_data_set
         https://stats.stackexchange.com/questions/74776/what-aspects-of-the-iris-data-set-make-it-so-successful-as-an-example-teaching/74901


b.	Formatting a data set
        https://gist.github.com/curran/a08a1080b88344b0c8a7
        https://archive.ics.uci.edu/ml/datasets/iris
        https://www.kaggle.com/jchen2186/machine-learning-with-iris-dataset

c.	Python code for statistics and graphical output
        https://docs.python.org/3/tutorial/introduction.html
        http://spronck.net/pythonbook/pythonbook.pdf
        https://www.youtube.com/watch?v=Nu7pGuenbwo
        https://www.youtube.com/watch?v=Nu7pGuenbwo
        https://www.youtube.com/watch?v=ywIWUfjPCyY
        https://datascience.stackexchange.com/questions/43057/how-to-classify-iris-flowers
        https://matplotlib.org/users/pyplot_tutorial.html